update tests with work from #2470
