+++ tests/infra/mock_fusillade_start.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+import time
+from mock_fusillade import MockFusilladeServer
+
+if __name__ == "__main__":
+    MockFusilladeServer.startServing()
